% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.model.R
\name{apply.model}
\alias{apply.model}
\title{Application of the CCC method on the real data}
\usage{
apply.model(peakTable, polarity)
}
\arguments{
\item{peakTable:}{the peakTable obtained from the runLC function of the metaMS package}

\item{polarity:}{the polarity of the MS experiment}
}
\value{
a list of grouped features with their isotopic pattern, estimated amount of Carbon and the CCC method predictions.
}
\description{
the function uses the peakTable given by metaMS to group the isotopic patterns and it uses this data to give an estimation of their amount of Carbons and predicts the CCC Y dependent variables on each of the isotopic pattern.
}
\examples{
data(peakTable)
model <- apply.model(peakTable, polarity = "negative")
#' 
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

