% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCC.xsaFA.R
\name{CCC.xsaFA}
\alias{CCC.xsaFA}
\title{Application of the CCC method on the xsaFA from CAMERA}
\usage{
CCC.xsaFA(xsaFA, models = NULL)
}
\arguments{
\item{xsaFA}{the xsaFA object resulting from CAMERA analysis}

\item{models}{the statistical models to use in the CCC_method. Default is NULL, so the original models are loaded and applied. Elsewhere, own models can be applied , building them using the function "model.building".}
}
\value{
a list of grouped features with their isotopic pattern, estimated amount of Carbon and the CCC method predictions.
}
\description{
the function uses the xsaFA given by CAMERA to group the isotopic patterns and it uses this data to give an estimation of their amount of Carbons and predicts the CCC Y dependent variables on each of the isotopic pattern. It can also handle statistical models built on different standards dataset using the function model.building.
}
\examples{
#do not run: try only with your own xsaFA object, it requires xcms and CAMERA
#tni <- CCC.xsaFA(xsaFA)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}
