% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metfrag_analysis.R
\name{metfragging}
\alias{metfragging}
\title{Online use of metfRag package to obtain MS/MS analysis of the markers using MSe data, starting from the tni list given by the function CCC.peakTable or CCC.xsaFA}
\usage{
metfragging(tni, ps_spectra, threshold, markers)
}
\arguments{
\item{tni}{the list of isotopic clusters obtained from the CCC.peakTable function or CCC.xsaFA function}

\item{ps_spectra}{the pseudo_spectra obtained from the ps_spec function}

\item{threshold}{the threhold to apply to use the fragment in the MS/MS analysis. default is 1000}

\item{markers}{a list of markers obtained from the analysis, indicates as row.name of the original peakTable/peaklist}
}
\value{
the scores given by metfRag
}
\description{
The function si a wrapper to use the run.metfrag function on all the markers using the TNI table and the CCC-method
}
\examples{
data(peakTable)
tni <- CCC.peakTable(peakTable, polarity "negative")
ps_spectra <- ps_spec(tni, peakTable)
scores <- metfragging(tni, ps_spectra)

}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

