% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.building.R
\name{model.building}
\alias{model.building}
\title{Build your own model for the CCC_method}
\usage{
model.building(X1Y, ny, models)
}
\arguments{
\item{X1Y}{the dataset built by the dataset building function}

\item{ny}{the Y dependent variable you want to build the model on (select a value between 1 and 9). Default option includes all the Y variables used in the CCC_method}

\item{models}{the statistical model you want to use to build the model. The options inclued logistic ("logistic"), lasso ("lasso"), ridge ("ridge"), pls ("pls"), and penalized-pls ("b_ppls"). The default option builds the models according to the CCC_method.}
}
\value{
a list of statistical models to perform the CCC approach
}
\description{
the function builds the statistical models to predict the Y dependent variables of theCCC_method  from a chemical standards dataset.
}
\examples{
STD_RP <- read.csv(system.file("extdata", "STD_RP.csv", package = "CCC"), row.names = 1, stringsAsFactors = F)
X1Y <- dataset.building(STD_RP)
models <- model.building(X1Y)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

