% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.testing.R
\name{model.testing}
\alias{model.testing}
\title{CCC_model testing for the evaluation of the standards dataset}
\usage{
model.testing(X1Y, ntest, ny, model, errori, per.test = FALSE)
}
\arguments{
\item{X1Y}{the dataset given by the dataset.building function}

\item{ntest}{the number of test to repeat on the dataset}

\item{ny}{the Y dependent variable under evaluation}

\item{model}{the statistical model to apply: choose between "pls", "lasso", "ridge", "logistic"}

\item{errori}{the kind of error you want to test on the dataset. Choose between "ppms", (mass accuracy error) "Cees" (Carbon estimation error), and "nerr" (normal distributed error)}

\item{per.test}{if TRUE it performs permutation test on the Y value, to determine the random level of relationship between X and Y. default is FALSE}
}
\value{
a list containing the values of the accuracy of the model across the error values. The accuracy is given with 3 parameters: the means_All (percentage of correct predictions) sds_All (percentage of standard deviation) and the RMSE (root mean squared error). It returns also a plot of the test, indicating the accuracy mean across the measurement error.
}
\description{
A wrapper function to test a specific model on the X-Y dataset generated by the dataset.building function
}
\examples{
STD_RP <- read.csv(system.file("extdata", "STD_RP.csv", package = "CCC"), row.names = 1, stringsAsFactors = F)
X1Y <- dataset.building(STD_RP)
results <- model.testing(X1Y, ntest=1000, ny=1, model = "logistic", errori = "nerr", per.test = F)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

