% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offline.CCC.R
\name{offline.CCC}
\alias{offline.CCC}
\title{manual application of the CCC method on data}
\usage{
offline.CCC(query = NULL, datapath, polarity, models = NULL)
}
\arguments{
\item{query}{a matrix data containing the RT, the mz of the main ions and their intensity}

\item{datapath}{the path to upload a csv files containing the RT, the mz of the main ions and their intensity}

\item{polarity}{the polarity of the experiment}

\item{models}{the statisticals models to apply for the prediction. If NULL, old models from previous CCC version will be uploaded}
}
\value{
a data.frame containing the predicted value for your specific query
}
\description{
The function uses the users data to predict the CCC Y dependent variables
}
\examples{
load(system.file("extdata", "models.rda", package = "CCC"))
load(system.file("extdata", "example1.rda", package = "CCC"))
offline.CCC(query = example1, polarity = "negative", models = models)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}
