% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_spec.R
\name{ps_spec}
\alias{ps_spec}
\title{Pseudospectra extraction for MSe data analysis with metfRag}
\usage{
ps_spec(tni, peakTable)
}
\arguments{
\item{"tni:}{the list of grouped features obtained from the function apply model"}

\item{"peakTable:}{the peakTable obtained from the runLC function from metaMS"}
}
\value{
"ps_spectra: all the pseudospectra per each group of feature from the list tni"
}
\description{
Pseudospectra extraction for MSe data analysis with metfRag
}
\examples{
ps_spectra <- ps_spec(tni, peakTable)
}

