% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.csv.R
\name{rank.csv}
\alias{rank.csv}
\title{Ranking function to get the best candidates from the metfRag analysis from .csv files}
\usage{
rank.csv(directory, tni)
}
\arguments{
\item{directory}{where the .csv files are stored. The file name must correspond to the feature number (rowname) in the original peaktable}

\item{tni}{is a list of grouped features given by the CCC.peakTable/CCC.xsaFA function}
}
\value{
a list of csv files reordered according to the CCC method parameters
}
\description{
The function reorders the metfrag candidates using the information given by the CCC_method. It uploads the .csv files from a directory, and performs reordering of the data inside each file. It returns a data.frame containing all the files together. NB: the name of the .csv file must correspond to the original raw number of the markers in the peakTable.
}
\examples{
data (peakTable)
tni <- apply.model(peakTable, polarity = "negative")
candidates <- rank.csv("H:/R/CCC/CCC_method/inst/extdata/CSV", tni)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

