% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.sdf.R
\name{rank.sdf}
\alias{rank.sdf}
\title{Ranking function to get the best candidates from the metfRag analysis from .sdf files}
\usage{
rank.sdf(directory, tni)
}
\arguments{
\item{directory}{where the sdf files are stored. The file name must correspond to the feature number (rowname) in the original peaktable}

\item{tni}{a list of grouped features given by the apply.model function}
}
\value{
a list of SDF files reordered according to the CCC method parameters
}
\description{
The function reorders the metfrag candidates using the information given by the CCC_method. It uploads the .sdf files from a directory, and performs reordering of the data inside each file. It returns a list of SDF files each containing a list of candidates. NB: the .sdf file names must correspond to the original raw number of the markers in the pekaTable.
}
\examples{
data (peakTable)
tni <- apply.model(peakTable, polarity = "negative")
candidates <- rank.sdf("H:/R/CCC/CCC_method/inst/extdata/SDF/", tni) 
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

