% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.2010.R
\name{ranking.2010}
\alias{ranking.2010}
\title{Ranking function to get the best candidates from the metfRag analysis from .csv files}
\usage{
ranking.2010(directory, tni)
}
\arguments{
\item{directory}{where the .csv files are stored. The file name must correspond to the feature number (rowname) in the original peaktable}

\item{tni}{is a list of grouped features given by the CCC.peakTable/CCC.xsaFA function}
}
\value{
a list of csv files reordered according to the CCC method parameters
}
\description{
The function reorders the metfrag candidates using the information given by the CCC_method. It uploads the .csv files from a directory, and performs reordering of the data inside each file. It returns a data.frame containing all the files together. NB: the .csv files must be from the web metfrag2010 and must be already converted in csv files (not xls). The name of the .csv file must correspond to the original raw number of the markers in the peakTable.
}
\examples{
data (peakTable)
tni <- CCC.peakTable(peakTable, polarity = "negative")
candidates <- ranking.2010(system.file("extdata", "2010", package = "CCC"), tni)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

