% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.2016.R
\name{ranking.2016}
\alias{ranking.2016}
\title{Ranking function to get the best candidates from the metfRag analysis from .sdf files}
\usage{
ranking.2016(directory, tni, sep)
}
\arguments{
\item{directory}{where the sdf files are stored. The file name must correspond to the feature number (rowname) in the original peaktable}

\item{tni}{a list of grouped features given by the apply.model function}

\item{sep}{is the separator preferred in the kind of file (useful only for csv files). Default is the "comma". Any separator can be applied.}
}
\value{
a list of SDF files reordered according to the CCC method parameters
}
\description{
The function reorders the metfrag candidates using the information given by the CCC_method. It uploads the .sdf files from a directory, and performs reordering of the data inside each file. It returns a list of SDF files each containing a list of candidates. NB: the .sdf file names must correspond to the original raw number of the markers in the pekaTable.
}
\examples{
data (peakTable)
tni <- CCC.peakTable(peakTable, polarity = "negative")
candidates <- ranking.2016(system.file("extdata", "2016", package = "CCC"), tni)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}
