% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.csv.R
\name{reorder.csv}
\alias{reorder.csv}
\title{Reordering function to get the best candidates from the metfRag analysis from .csv files}
\usage{
\method{reorder}{csv}(directory, compi)
}
\arguments{
\item{directory}{where the .csv files are stored}

\item{compi:}{a data.frame containing all the features extracted by the apply.model function}
}
\value{
a list of csv files reordered according to the CCC method parameters
}
\description{
Reordering function to get the best candidates from the metfRag analysis from .csv files
}

