% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.sdf.R
\name{reorder.sdf}
\alias{reorder.sdf}
\title{Reordering function to get the best candidates from the metfRag analysis from .sdf files}
\usage{
\method{reorder}{sdf}(directory, tni)
}
\arguments{
\item{directory}{where the sdf files are stored. The file name must correspond to the feature number (rowname) in the original peaktable}

\item{tni:}{a list of grouped features given by the apply.model function}
}
\value{
a list of SDF files reordered according to the CCC method parameters
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

