% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.testing.R, R/test.model.R
\name{test.model}
\alias{test.model}
\alias{test.model}
\title{CCC_model testing for the evaluation of the standards dataset}
\usage{
test.model(X1Y, model, variables, ntest, errors)

test.model(X1Y, model, variables, ntest, errors)
}
\arguments{
\item{X1Y}{the dataset given by the dataset.building function}

\item{model}{the statistical model to apply: choose between "pls", "lasso", "ridge", "logistic"}

\item{variables}{chose eiher between the "categorical" or "ordinal" variables. Categorical variables include: Sulfur, Phosphate, acidic, Nitrogen, Heterocycles, C/O ratio and Glycosides. The ordinal variables include the phenols and aliphatic moiety.}

\item{ntest}{the number of test to repeat on the dataset}

\item{errors}{the kind of error you want to test on the dataset. Choose between "ppm", (mass accuracy error) "Cees" (Carbon estimation error), and "nerr" (normal distributed error)}

\item{X1Y}{the dataset given by the dataset.building function}

\item{ntest}{the number of test to repeat on the dataset}

\item{variables}{chose eiher between the "categorical" or "ordinal" variables. Categorical variables include: Sulfur, Phosphate, acidic, Nitrogen, Heterocycles, C/O ratio and Glycosides. The ordinal variables include the phenols and aliphatic moiety.}

\item{model}{the statistical model to apply: choose between "pls", "lasso", "ridge", "logistic"}

\item{errors}{the kind of error you want to test on the dataset. Choose between "ppm", (mass accuracy error) "Cees" (Carbon estimation error), and "nerr" (normal distributed error)}
}
\value{
a list containing the values of the accuracy of the model across the error values. The accuracy is given with 3 parameters: the means_All (percentage of correct predictions) sds_All (percentage of standard deviation) and the MSEE (root mean squared error). It returns also a plot of the test, indicating the accuracy mean across the measurement error.

a list containing the values of the accuracy of the model across the error values. The accuracy is given with 3 parameters: the means_All (percentage of correct predictions) sds_All (percentage of standard deviation) and the MSEE (root mean squared error). It returns also a plot of the test, indicating the accuracy mean across the measurement error.
}
\description{
A wrapper function to test a specific model on the X-Y dataset generated by the dataset.building function

A wrapper function to test a specific model on the X-Y dataset generated by the dataset.building function
}
\examples{
STD_RP <- read.csv(system.file("extdata", "STD_RP.csv", package = "CCC"), row.names = 1, stringsAsFactors = FALSE)
X1Y <- dataset.building(STD_RP)
results <- test.model(X1Y, model = "lasso", variables, ntest = 100, errors = "ppm")
STD_RP <- read.csv(system.file("extdata", "STD_RP.csv", package = "CCC"), row.names = 1, stringsAsFactors = FALSE)
X1Y <- dataset.building(STD_RP)
results <- test.model(X1Y, model = "lasso", variables = "categorical", ntest = 100, errors = "ppm")
}
\author{
Luca Narduzzi "nardluca@gmail.com"

Luca Narduzzi "nardluca@gmail.com"
}
