% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.building.R
\name{dataset.building}
\alias{dataset.building}
\title{Chemical standards dataset building for the construction of the CCC method dataset}
\usage{
dataset.building(x)
}
\arguments{
\item{x:}{a data.frame containing the database of chemical standards of your laboratory}
}
\value{
the basic dataset to build a CCC_method
}
\description{
Chemical standards dataset building for the construction of the CCC method dataset
}
\examples{
X1Y <- dataset.building(STD_RP)
}

