% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.csv.R
\name{reorder.csv}
\alias{reorder.csv}
\title{Reordering function to get the best candidates from the metfRag analysis from .csv files}
\usage{
\method{reorder}{csv}(directory, tni)
}
\arguments{
\item{directory}{where the .csv files are stored. The file name must correspond to the feature number (rowname) in the original peaktable}

\item{tni:}{a list of grouped features given by the apply.model function}
}
\value{
a list of csv files reordered according to the CCC method parameters
}
\description{
Reordering function to get the best candidates from the metfRag analysis from .csv files
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}

