% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.sdf.R
\name{reorder.sdf}
\alias{reorder.sdf}
\title{Reordering function to get the best candidates from the metfRag analysis from .sdf files}
\usage{
\method{reorder}{sdf}(directory, compi)
}
\arguments{
\item{directory}{where the sdf files are stored}

\item{compi:}{a data.frame containing all the features extracted by the apply.model function}
}
\value{
a list of SDF files reordered according to the CCC method parameters
}
\description{
Reordering function to get the best candidates from the metfRag analysis from .sdf files
}

